Fix #108: make test that failed with ICU 76 more robust
This removes the timezone string which apparently has changed from
"GMT+1" (ICU < 76) to "MEZ" (ICU 76) (German rendering).

I anyway do not understand why the timezone is printed in the
`LongFormatStyle`, because the ICU documentation suggests it should
only be printed in the `FullFormatStyle`.

https://github.com/unicode-org/icu/blob/72406ed78a95eec2be3d19ef7e7ce2c2cca7dd5e/icu4c/source/i18n/unicode/udat.h#L128-L130

 *   <li>   UDAT_LONG is longer, such as January 12, 1952 or 3:30:32pm
 *   <li>   UDAT_FULL is pretty completely specified, such as
 *          Tuesday, April 12, 1952 AD or 3:30:42pm PST.

Anyway switching to `MediumFormatStyle` solves the problem at hand.